+++ frontend/src/admin-performance/components/TrendChart.vue
@@ -1,3 +1,0 @@
-<script>
-export { default } from '../../admin-shared/components/TrendChart.vue';
-</script>